apps/voice/requirements.txt
livekit-agents==1.2.4
python-dotenv==1.0.1
aiohttp==3.10.5

